requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# LangGraph and LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-groq==0.2.1

# Content Extraction
pytesseract==0.3.10
pdf2image==1.17.0
PyPDF2==3.0.1
Pillow==10.2.0
openai-whisper==20231117
pydub==0.25.1

# YouTube
youtube-transcript-api==0.6.2

# Utilities
python-dotenv==1.0.0
pydantic
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Frontend (Optional - Streamlit)
streamlit==1.31.0